requirements.txt
numpy==1.23.5
wandb
ipdb
pybullet==3.2.5
h5py
opencv-python
pytorch-lightning
tqdm
pre-commit
black
scipy
imageio
open3d
git+https://github.com/fishbotics/atob@v0.0.1
pyquaternion
hydra-core
imageio[ffmpeg]
geometrout==0.0.3.4
trimesh
yourdfpy
shapely
# diffusion policy reqs: #not clear which of those belonged exclusively to the diffusion policy
dill==0.3.5.1
einops==0.7.0
diffusers==0.11.1
zarr==2.12.0
numba==0.56.4
av==10.0.0
wheel==0.38.4
setuptools==65.6.3
gym==0.21.0
imagecodecs
submitit
plotly
stable_baselines3
shimmy
pycollada
meshcat
transformers==4.34.0
objaverse
numpy-quaternion
moviepy
kornia
pupil_apriltags
huggingface_hub
